+++ frontend/src/Components/Reglage/Reglage.jsx
@@ -1,0 +1,23 @@
+import React from 'react'
+import './Reglage'
+import '../../App.css'
+import Sidebar from '../Sidebar/Sidebar';
+import Topbar from '../Topbar/Topbar';
+
+
+const Reglage = () => {
+
+
+    return (
+
+        <div className="App">
+            <div className="AppGlass">
+                <Sidebar />
+                <Topbar />
+            </div>
+        </div>
+
+    )
+}
+
+export default Reglage
