+++ frontend/src/Components/Sidebar/Sidebar.css
@@ -1,5 +1,4 @@
 .Sidebar {
-    display: flex;
     background-color: rgba(255, 255, 255, 0.7);
     height: 100vh;
 }
