--- frontend/src/Components/Sidebar/Sidebar.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import './Sidebar.css'
-
-const Sidebar = () => {
-    return (
-        <div className='Sidebar'>
-
-            <div className='Logo'>
-                <span>
-                    DALY <span>PAPER.</span>
-                </span>
-            </div>
-        </div>
-    )
-}
-
-export default Sidebar
